complete SEND_REPORT and add CREATE_RECORDS webhook notification
